Merge pull requests on successful commit statuses
Basically mimics the `LGTM on green`.

When a commit status event comes through with the `success` state the
bot will go through and attempt to see if any PR's with that specific
SHA should be merged.

Also does checks like if the HEAD sha is the same as the SHA that the
commit status is for, and won't continue if there is a mismatch.

Quite a bit of refactoring in terms of abstracting away the
mergePullRequest method to allow for it to be used in both the context
of commit statuses as well as pull request review events.

Signed-off-by: Eli Uriegas <eli.uriegas@docker.com>